refactor: move download tracking fns into file
